requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
motor==3.6.0
pydantic==2.9.2
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic-settings
Pillow
python-multipart